api_gateway/pyproject.toml
[project]
name = "api_gateway"
version = "0.1.0"
description = "..."
authors = [{name = "Jahongir", email = "jahongir.hodjaniyazov@gmail.com"}]
requires-python = ">=3.13"
dependencies = [
    "fastapi ==0.116.1",
    "pydantic ==2.11.7",
    "uvicorn ==0.35.0",
    "pydantic-settings ==2.10.1",
    "aiokafka ==0.12.0",
    "httpx ==0.28.1",
    "python-jose ==3.5.0",
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
